Reverting changes to importing styles
In an earlier push the style for importing was changed to use explicit
relative imports. Should have tested it out better but it's currently
not working due to the way that q2mm is set up. To get q2mm packaged
more properly, some work is needed to change the way q2mm is called
and run, and it needs to be packaged more like a real python package.
In the mean time using the old style of imports is still fine until
this all gets restructured a little. Also with the aim of packaging
q2mm properly, I've started a requirements.txt file which should contain
other packages required to run q2mm which may not come with all
python distributions by default.